fix(harness): reset exempt progress streaks
